+++ onediff_sd_webui_extensions/compile/oneflow/utils.py
@@ -2,10 +2,23 @@ from onediff_utils import singleton_decorator
 
 from onediff.infer_compiler.backends.oneflow.transform import register
 
-from .mock import ldm, sgm
-
 
 @singleton_decorator
 def init_oneflow_backend():
+    try:
+        import oneflow as flow
+    except ImportError:
+        raise RuntimeError(
+            "Backend oneflow for OneDiff is invalid, please make sure you have installed OneFlow"
+        )
+
+    from .mock import controlnet, ldm, sgm
+
     register(package_names=["ldm"], torch2oflow_class_map=ldm.torch2oflow_class_map)
     register(package_names=["sgm"], torch2oflow_class_map=sgm.torch2oflow_class_map)
+    register(
+        package_names=["scripts.hook"],
+        torch2oflow_class_map=controlnet.torch2oflow_class_map,
+    )
+
+
