+++ onediff_sd_webui_extensions/onediff_controlnet/__init__.py
@@ -1,0 +1,6 @@
+from .compile import onediff_controlnet_decorator, compile_controlnet_ldm_unet
+from .hijack import hijack_controlnet_extension
+
+__all__ = [
+    "onediff_controlnet_decorator",
+]
