+++ onediff_sd_webui_extensions/onediff_shared.py
@@ -9,6 +9,7 @@ previous_unet_type = {
     "is_ssd": False,
 }
 onediff_enabled = False
+onediff_backend = "oneflow"
 
 # controlnet
 controlnet_enabled = False
